Intégration graphismes astéroïdes, station, vaisseaux. Net Worth prend maintenant en compte la valeur des vaisseaux et de leur cargo.

+++ LD45 Project/Assets/Sources/UI/GeneralUIFigures.cs
@@ -14,8 +14,8 @@ public class GeneralUIFigures : MonoBehaviour
 
     private void Update()
     {
-        playerNetWorthText.text = "$ " + GameManager.NetWorth.ToString();
-        playerCashText.text = "$ " + GameManager.PlayerCash;
-        playerDebtText.text = "$ " + GameManager.Debt;
+        playerNetWorthText.text = "$ " + ((int)(GameManager.NetWorth)).ToString();
+        playerCashText.text = "$ " + (int)GameManager.PlayerCash;
+        playerDebtText.text = "$ " + (int)GameManager.Debt;
     }
 }
